+++ src/main/java/collection/Cars.java
@@ -1,4 +1,4 @@
-package domain;
+package collection;
 
 import java.util.ArrayList;
 import java.util.List;
