+++ src/test/java/domain/CarsTest.java
@@ -10,6 +10,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import collection.Car;
+import collection.Cars;
 import constant.ConstantNumbers;
 
 class CarsTest {
